A pass of renames and fixes for frame graph nodes

+++ oblo/renderer/include/oblo/renderer/graph/frame_graph_resources.hpp
@@ -19,7 +19,7 @@ namespace oblo
 
     struct buffer_resource_initializer
     {
-        u32 size{};
+        u64 size{};
         std::span<const byte> data;
         bool isStable{};
     };
